refactor: split editor window into modular UI
