+++ editor/__init__.py
@@ -1,4 +1,6 @@
-from .main_editor import EditorWindow, Canvas, ColorButton
+from .editor_window import EditorWindow
+from .ui.canvas import Canvas
+from .ui.color_widgets import ColorButton
 from .editor_logic import EditorLogic
 
 __all__ = ['EditorWindow', 'Canvas', 'ColorButton', 'EditorLogic']
